+++ src/future/enums.ts
@@ -1,0 +1,36 @@
+export enum CHANNEL {
+	AI1 = 1,
+	AI2 = 2,
+	AI3 = 3,
+	AI4 = 4,
+	AI5 = 5,
+	AI6 = 6,
+	AX1 = 7,
+	AX2 = 8
+}
+
+export enum COMMUNICATION_MODE {
+	BLUETOOTH,
+	WEB_SOCKET
+}
+
+export enum ADC_UNIT {
+	ADC_UNIT_1 = 1,
+	ADC_UNIT_2 = 2,
+	ADC_UNIT_BOTH = 3,
+	ADC_UNIT_ALTER = 7
+}
+
+export enum ADC_ATTEN {
+	ADC_ATTEN_DB_0 = 0,
+	ADC_ATTEN_DB_2_5 = 1,
+	ADC_ATTEN_DB_6 = 2,
+	ADC_ATTEN_DB_11 = 3
+}
+
+export enum ADC_BITS_WIDTH {
+	ADC_BITS_WIDTH_9 = 1,
+	ADC_BITS_WIDTH_10 = 2,
+	ADC_BITS_WIDTH_11 = 3,
+	ADC_BITS_WIDTH_12 = 4
+}
